+++ src/components/index.js
@@ -1,3 +1,4 @@
 
 export {CustomNavBar} from "./navbar";
 export {CustomFooter} from "./footer"
+export {CartCell} from "./cart-cell"
