+++ src/components/navbar/index.js
@@ -1,4 +1,4 @@
-import { Col, Form, Row, Button } from 'react-bootstrap';
+import { Form, Button } from 'react-bootstrap';
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
