+++ src/controllers/cart/index.js
@@ -1,94 +1,15 @@
-import { Button, Col, Container, Form, Image, Row } from "react-bootstrap"
+import { Button, Col, Container, Row } from "react-bootstrap"
 import { useDispatch, useSelector } from "react-redux"
-import Card from 'react-bootstrap/Card';
-import "./index.css";
 import { useEffect, useState } from "react";
 import useAPI from "../../api";
 import { useNavigate } from "react-router-dom";
-
-const CartCell2 = ({ item, deleteF, onNumberChange }) => {
-    let options = [];
-    item.car.attributeTypes.forEach(type => {
-        type.items.filter(o => o.selected === true).forEach(i => {
-            options.push(i);
-        })
-    });
-
-    const numberChange = (e) => {
-        e.preventDefault();
-        const value = parseInt(e.target.value);
-        const unitPrice = item.totalPrice /item.number;
-
-        if (value) {
-            let temp = {...item, number: value, totalPrice: unitPrice * value}
-            onNumberChange(item, temp);
-        } 
-    }
-
-    const handleDelete = (e) => {
-        e.preventDefault();
-        deleteF(item);
-    }
-
-    return (
-        <Container className="cartitem">
-            <Row >
-                <Col lg={3}>
-                    <Image rounded className="w-100" src={item.car.images[0]} />
-                </Col>
-                <Col lg={6}>
-                    <div>
-                        <Card.Title>{item.car.name.toUpperCase()}</Card.Title>
-                        <Card.Text>
-                            Total Price: <span className="text-success">${item.totalPrice}</span>
-                        </Card.Text>
-
-                        <Card.Text>
-                            Make: {item.car.make}
-                        </Card.Text>
-                        <Card.Text>
-                            Model: {item.car.model}
-                        </Card.Text>
-                    </div>
-
-                    <div className="text-center d-flex">
-                        {options.map(e => {
-                            return <div className="cart-cell" key={e.value} >
-                                <p className=" text-uppercase">
-                                    {e.value}
-                                </p>
-                                <p className={e.selected ? "" : " text-success"}>${e.additionalPrice}</p>
-                            </div>
-                        })}
-                    </div>
-
-
-
-                </Col>
-                <Col lg={3}>
-                    <div className="mb-6 mb-lg-0 text-center cart-select"  >
-                        <Form.Select onChange={numberChange} value={item.number} aria-label="Default select example">
-                            <option value="1">1</option>
-                            <option value="2">2</option>
-                            <option value="3">3</option>
-                            <option value="4">4</option>
-                        </Form.Select>
-                    </div>
-                    <div className="button-box">
-                        <Button value={item.car.productNumber} rounded variant="danger" onClick={handleDelete} >Delete</Button>
-
-                    </div>
-                </Col>
-            </Row>
-
-        </Container>
-
-    )
-}
-
+import { CartCell } from "../../components";
+import "./index.css";
 
 export const Cart = () => {
     const carts = useSelector(state => state.carts);
+    const user = useSelector(state => state.user);
+
     const dispatch = useDispatch();
 
     const [total, setTotal] = useState(0);
@@ -97,9 +18,10 @@ export const Cart = () => {
 
     const checkout = async (e) => {
         e.preventDefault();
-        const response = await PostClient("/api/orders",carts);
+        const order = {items: carts, email: user.email, orderStatus: "PLACED", total: total};
+        const response = await PostClient("/api/orders",order);
         if (response.status === 200) {
-            console.log("status 200");
+            console.log(response.data);
             navigate("/orders");
         }
     }
@@ -128,7 +50,7 @@ export const Cart = () => {
             <Row>
                 {carts.map(e => (
                     <Col lg={12} key={e.car.productNumber}>
-                        <CartCell2 item={e} deleteF={deleteF} onNumberChange={onNumberChange}/>
+                        <CartCell item={e} deleteF={deleteF} onNumberChange={onNumberChange}/>
                     </Col>
                 ))}
             </Row>
